# src/pydas/static/swagger-config.yaml
swagger: "2.0"
info:
  description: "API for accessing and interacting with sDAS utility and metadata."
  version: "1.0.0"
  title: "stock Data Acquisition Service (sDAS) API"
host: "localhost:5000"
basePath: "/api/v1"
...
schemes:
  - "http"
securityDefinitions:
  AuthorizationHeader:
    type: apiKey
    in: header
    name: Authorization
security:
  - AuthorizationHeader: []
paths:
...
  /companies/{company_symbol}/features/{feature_name}:
    get:
      tags:
        - "company features"
      summary: "Retrieves a single registered feature"
      description: ""
      operationId: "getCompanyFeature"
      parameters:
        - name: "company_symbol"
          in: "path"
          description: "Symbol of company to retrieve features for"
          required: true
          type: string
        - name: "feature_name"
          in: "path"
          description: "Name of feature to return"
          required: true
          type: string
      produces:
        - "application/json"
      responses:
        200:
          description: "Feature object registered in the system with a matching name"
        401:
          description: "Invalid or null Authorization header (apiKey)"
        403:
          description: "User is not authorized to access this data"
        404:
          description: "Company or feature not found"
    patch:
      tags:
        - "company features"
      summary: "Update an existing feature registry"
      description: ""
      operationId: "updateCompanyFeature"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "company_symbol"
          in: "path"
          description: "Symbol of company to retrieve features for"
          required: true
          type: string
        - name: "feature_name"
          in: "path"
          description: "Name of feature to update"
          required: true
          type: string
        - in: "body"
          name: "body"
          description: "Feature object to update in the system"
          required: true
          schema:
            $ref: "#/definitions/Feature"
      responses:
        200:
          description: "Updated feature object"
        400:
          description: "Invalid feature object supplied"
        401:
          description: "Invalid or null Authorization header (apiKey)"
        403:
          description: "User is not authorized to access this data"
        404:
          description: "Company or feature not found"
    delete:
      tags:
        - "company features"
      summary: "Deletes a feature"
      description: ""
      operationId: "deleteCompanyFeature"
      parameters:
        - name: "company_symbol"
          in: "path"
          description: "Symbol of company to retrieve features for"
          required: true
          type: string
        - name: "feature_name"
          in: "path"
          description: "Name of feature to delete"
          required: true
          type: string
      responses:
        204:
          description: "Feature successfully deleted"
        401:
          description: "Invalid or null Authorization header (apiKey)"
        403:
          description: "User is not authorized to access this data"
        404:
          description: "Company or feature not found"
  /features:
    get:
      tags:
        - "features"
      summary: "Retrieves all registered features"
      description: ""
      operationId: "getFeatures"
      produces:
        - "application/json"
      responses:
        200:
          description: "Array of feature objects registered in the system"
        401:
          description: "Invalid or null Authorization header (apiKey)"
        403:
          description: "User is not authorized to access this data"
    post:
      tags:
        - "features"
      summary: "Registers a new feature in the system"
      description: ""
      operationId: "addFeature"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Feature object that needs to be registered"
          required: true
          schema:
            $ref: "#/definitions/Feature"
      responses:
        201:
          description: "Newly registered feature"
        400:
          description: "Invalid input"
        401:
          description: "Invalid or null Authorization header (apiKey)"
        403:
          description: "User is not authorized to access this data"
  /features/{feature_name}:
    get:
      tags:
        - "features"
      summary: "Retrieves a single registered feature"
      description: ""
      operationId: "getFeature"
      parameters:
        - name: "feature_name"
          in: "path"
          description: "Name of feature to return"
          required: true
          type: string
      produces:
        - "application/json"
      responses:
        200:
          description: "Feature object registered in the system with a matching name"
        401:
          description: "Invalid or null Authorization header (apiKey)"
        403:
          description: "User is not authorized to access this data"
        404:
          description: "Feature not found"
    patch:
      tags:
        - "features"
      summary: "Update an existing feature registry"
      description: ""
      operationId: "updateFeature"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "feature_name"
          in: "path"
          description: "Name of feature to update"
          required: true
          type: string
        - in: "body"
          name: "body"
          description: "feature object to update in the system"
          required: true
          schema:
            $ref: "#/definitions/Feature"
      responses:
        200:
          description: "Updated feature object"
        400:
          description: "Invalid feature object supplied"
        401:
          description: "Invalid or null Authorization header (apiKey)"
        403:
          description: "User is not authorized to access this data"
        404:
          description: "Feature not found"
    delete:
      tags:
        - "features"
      summary: "Deletes a feature"
      description: ""
      operationId: "deleteFeature"
      parameters:
        - name: "feature_name"
          in: "path"
          description: "Name of feature to delete"
          required: true
          type: string
      responses:
        204:
          description: "Feature successfully deleted"
        401:
          description: "Invalid or null Authorization header (apiKey)"
        403:
          description: "User is not authorized to access this data"
        404:
          description: "Feature not found"
...
  /statistics/company/{company_symbol}/features/{feature_name}:
    get:
      tags:
        - "statistics"
      summary: "Retrieves system statistics for a given company feature"
      description: ""
      operationId: "getCompanyFeatureStatistics"
      produces:
        - "application/json"
      parameters:
        - name: "company_symbol"
          in: "path"
          description: "Company symbol to return statistics for"
          required: true
          type: string
        - name: "feature_name"
          in: "path"
          description: "Feature name to return statistics for"
          required: true
          type: string
      responses:
        200:
          description: "Feature statistics for the given company"
          schema:
            $ref: "#/definitions/Statistics"
        401:
          description: "Invalid or null Authorization header (apiKey)"
        403:
          description: "User is not authorized to access this data"
definitions:
...
  Feature:
    type: object
    description: "Metadata for controlling where and how to acquire data"
    required:
      - "name"
      - "uri"
      - "handler"
    properties:
      name:
        type: string
        example: "open"
      uri:
        type: string
        description: "third-party API endpoint to acquire the data from"
        example: "/stock/{:symbol}/chart"
      description:
        type: string
        example: "Price of stock at start of trading day"
      handler:
        $ref: "#/definitions/Handler"
  Handler:
    type: object
    description: "Extensible function for acquiring data from third-party"
    required:
      - "id"
    properties:
      id:
        type: integer
        format: int32
        example: 1
      name:
        type: string
        description: "Function name to call to acquire data"
        example: "batch_handler"
...
  Statistics:
    type: object
    required:
      - "company_symbol"
      - "total_row_count"
    properties:
      company_symbol:
        type: string
        example: "GOOGL"
      feature_name:
        type: string
        example: "open"
      retrieval_datetime:
        type: string
        format: date-time
      total_row_count:
        type: integer
        format: int32